pyproject.toml
[tool.poetry]
name = "sa4e-ws2425-goichmann"
version = "0.1.0"
description = "Repository für Übungscode für das Modul Softwarearchitectures for Enterprises an der Universität Trier, Wintersemester 24/25"
authors = ["Michael-Goichmann <m.goichmann@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
setproctitle = "^1.3.4"
grpcio = "^1.68.0"
grpcio-tools = "^1.68.0"
protobuf = "^5.29.0"
numpy = "^2.1.3"
flask = "^3.1.0"
pymongo = "^4.10.1"
locust = "^2.32.6"
kafka-python = "^2.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

